+++ gobigger/agents/__init__.py
@@ -1,4 +1,2 @@
 from .base_agent import BaseAgent
 from .bot_agent import BotAgent
-from .smart_agent import SmartAgent
-from .level_agent import LevelBotAgent
